+++ modules/utils/defaultUtils.js
@@ -57,7 +57,9 @@ export const defaultRuleProperties = (config, parentRuleGroupPath = null, item =
   }
   
   if (field && operator) {
-    let {newValue, newValueSrc, newValueType, newValueError} = getNewValueForFieldOp(config, config, current, field, operator, "operator", false);
+    let {newValue, newValueSrc, newValueType, newValueError} = getNewValueForFieldOp(
+      config, config, current, field, operator, "operator", false
+    );
     current = current
       .set("value", newValue)
       .set("valueSrc", newValueSrc)
@@ -102,10 +104,6 @@ export const defaultRule = (id, config) => ({
 });
 
 export const defaultRoot = (config) => {
-  if (config.tree) {
-    return new Immutable.Map(config.tree);
-  }
-  
   return new Immutable.Map({
     type: "group",
     id: uuid(),