+++ tests/specs/WidgetsVanilla.test.js
@@ -4,8 +4,8 @@ import { with_qb, with_qb_skins } from "../support/utils";
 
 
 describe("vanilla widgets interactions", () => {
-  it("change number value", () => {
-    with_qb_skins(configs.with_all_types, inits.with_number, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change number value", async () => {
+    await with_qb_skins(configs.with_all_types, inits.with_number, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget input")
         .simulate("change", { target: { value: "3" } });
@@ -15,8 +15,8 @@ describe("vanilla widgets interactions", () => {
     });
   });
 
-  it("change text value", () => {
-    with_qb_skins(configs.with_all_types, inits.with_text, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change text value", async () => {
+    await with_qb_skins(configs.with_all_types, inits.with_text, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget input")
         .simulate("change", { target: { value: "def" } });
@@ -26,8 +26,8 @@ describe("vanilla widgets interactions", () => {
     });
   });
 
-  it("change date value", () => {
-    with_qb(configs.with_all_types, inits.with_date, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change date value", async () => {
+    await with_qb(configs.with_all_types, inits.with_date, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget input")
         .simulate("change", { target: { value: "2020-05-05" } });
@@ -37,8 +37,8 @@ describe("vanilla widgets interactions", () => {
     });
   });
 
-  it("change datetime value", () => {
-    with_qb(configs.with_all_types, inits.with_datetime, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change datetime value", async () => {
+    await with_qb(configs.with_all_types, inits.with_datetime, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget input")
         .simulate("change", { target: { value: "2020-05-05T02:30" } });
@@ -48,8 +48,8 @@ describe("vanilla widgets interactions", () => {
     });
   });
 
-  it("change select value", () => {
-    with_qb(configs.with_all_types, inits.with_select, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change select value", async () => {
+    await with_qb(configs.with_all_types, inits.with_select, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget select")
         .simulate("change", { target: { value: "green" } });
@@ -59,8 +59,8 @@ describe("vanilla widgets interactions", () => {
     });
   });
 
-  it("change multiselect value", () => {
-    with_qb(configs.with_all_types, inits.with_multiselect, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change multiselect value", async () => {
+    await with_qb(configs.with_all_types, inits.with_multiselect, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget select")
         .simulate("change", { target: { options: [ {value: "yellow", selected: true}, {value: "green"}, {value: "orange"} ] } });
@@ -77,8 +77,8 @@ describe("vanilla widgets interactions", () => {
     });
   });
 
-  it("change bool value", () => {
-    with_qb(configs.with_all_types, inits.with_bool, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change bool value", async () => {
+    await with_qb(configs.with_all_types, inits.with_bool, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget input[value=false]")
         .simulate("change", { target: { value: "false" } });
@@ -88,8 +88,8 @@ describe("vanilla widgets interactions", () => {
     });
   });
 
-  it("change slider value", () => {
-    with_qb(configs.with_all_types, inits.with_slider, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change slider value", async () => {
+    await with_qb(configs.with_all_types, inits.with_slider, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget input[type='range']")
         .simulate("change", { target: { value: 42 } });
@@ -99,8 +99,8 @@ describe("vanilla widgets interactions", () => {
     });
   });
 
-  it("change time value", () => {
-    with_qb(configs.with_all_types, inits.with_time, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change time value", async () => {
+    await with_qb(configs.with_all_types, inits.with_time, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--widget input[type='time']")
         .simulate("change", { target: { value: "10:30" } });